I did sort with Comparator

+++ src/main/java/com/gergert/task1/entity/ArrayData.java
@@ -1,4 +1,4 @@
 package com.gergert.task1.entity;
 
-public record ArrayData (int max, int min, int avg, int sum){
+public record ArrayData (int max, int min, double avg, int sum){
 }
